# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DocScan — Document Scan & Email System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  user: 'user',
  admin: 'admin',
} as const;

export type UserStatus = typeof UserStatus[keyof typeof UserStatus];


export const UserStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  status: UserStatus;
  createdAt: string;
  /** @nullable */
  documentCount?: number | null;
  /** @nullable */
  lastActivity?: string | null;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: UserInputRole;
}

export type UserUpdateStatus = typeof UserUpdateStatus[keyof typeof UserUpdateStatus];


export const UserUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  user: 'user',
  admin: 'admin',
} as const;

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  status?: UserUpdateStatus;
  /** @minLength 6 */
  password?: string;
  role?: UserUpdateRole;
}

export interface Document {
  id: number;
  userId: number;
  fileName: string;
  filePath?: string;
  fileType: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
}

export type EmailLogStatus = typeof EmailLogStatus[keyof typeof EmailLogStatus];


export const EmailLogStatus = {
  queued: 'queued',
  sent: 'sent',
  failed: 'failed',
} as const;

export interface EmailLog {
  id: number;
  documentId: number;
  senderId: number;
  recipientEmail: string;
  status: EmailLogStatus;
  /** @nullable */
  sentAt: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  documentName?: string | null;
  /** @nullable */
  senderName?: string | null;
}

export interface DocumentWithLogs {
  id: number;
  userId: number;
  fileName: string;
  filePath?: string;
  fileType: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
  emailLogs: EmailLog[];
}

export interface DocumentWithUser {
  id: number;
  userId: number;
  fileName: string;
  filePath?: string;
  fileType: string;
  /** @nullable */
  fileSize?: number | null;
  uploadedAt: string;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  userEmail?: string | null;
  emailLogs?: EmailLog[];
}

export interface SendResult {
  success: boolean;
  message?: string;
  logs: EmailLog[];
}

export interface Recipient {
  id: number;
  /** @nullable */
  userId?: number | null;
  recipientEmail: string;
  createdAt: string;
}

export interface RecipientInput {
  recipientEmail: string;
  /** @nullable */
  userId?: number | null;
}

export interface Settings {
  id: number;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: number | null;
  /** @nullable */
  smtpUser?: string | null;
  /** @nullable */
  smtpPass?: string | null;
  maxRecipients: number;
  maxFileSizeMb: number;
  /** @nullable */
  allowedFileTypes?: string | null;
}

export interface SettingsInput {
  smtpHost?: string;
  smtpPort?: number;
  smtpUser?: string;
  smtpPass?: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  maxRecipients?: number;
  /** @minimum 1 */
  maxFileSizeMb?: number;
  allowedFileTypes?: string;
}

export interface AuditLog {
  id: number;
  action: string;
  /** @nullable */
  userId: number | null;
  /** @nullable */
  userName?: string | null;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalUsers: number;
  activeUsers: number;
  totalDocuments: number;
  totalEmailsSent: number;
  failedEmails: number;
  documentsToday?: number;
  emailsToday?: number;
  recentActivity: AuditLog[];
}