# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DocScan — Document Scan & Email System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin-users]
      summary: List all users
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [admin-users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/users/{id}:
    patch:
      operationId: updateUser
      tags: [admin-users]
      summary: Update user (activate/deactivate/reset password)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteUser
      tags: [admin-users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/documents/{id}:
    delete:
      operationId: deleteDocument
      tags: [admin-documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Document deleted
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/recipients:
    get:
      operationId: listRecipients
      tags: [admin-recipients]
      summary: List configured recipient addresses
      responses:
        "200":
          description: Recipients list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recipient"
    post:
      operationId: createRecipient
      tags: [admin-recipients]
      summary: Add a recipient address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecipientInput"
      responses:
        "201":
          description: Recipient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipient"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/recipients/{id}:
    delete:
      operationId: deleteRecipient
      tags: [admin-recipients]
      summary: Remove a recipient address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Recipient deleted
        "404":
          description: Recipient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin-settings]
      summary: Get system settings
      responses:
        "200":
          description: System settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
    put:
      operationId: updateSettings
      tags: [admin-settings]
      summary: Update system settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
          format: date-time
        documentCount:
          type: ["integer", "null"]
        lastActivity:
          type: ["string", "null"]
    UserInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [user, admin]
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        status:
          type: string
          enum: [active, inactive]
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [user, admin]
...
    Recipient:
      type: object
      required: [id, recipientEmail, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        recipientEmail:
          type: string
        createdAt:
          type: string
          format: date-time
    RecipientInput:
      type: object
      required: [recipientEmail]
      properties:
        recipientEmail:
          type: string
        userId:
          type: ["integer", "null"]
    Settings:
      type: object
      required: [id, maxRecipients, maxFileSizeMb]
      properties:
        id:
          type: integer
        smtpHost:
          type: ["string", "null"]
        smtpPort:
          type: ["integer", "null"]
        smtpUser:
          type: ["string", "null"]
        smtpPass:
          type: ["string", "null"]
        maxRecipients:
          type: integer
        maxFileSizeMb:
          type: integer
        allowedFileTypes:
          type: ["string", "null"]
    SettingsInput:
      type: object
      properties:
        smtpHost:
          type: string
        smtpPort:
          type: integer
        smtpUser:
          type: string
        smtpPass:
          type: string
        maxRecipients:
          type: integer
          minimum: 1
          maximum: 5
        maxFileSizeMb:
          type: integer
          minimum: 1
        allowedFileTypes:
          type: string
...