# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * DocScan — Document Scan & Email System API
 * OpenAPI spec version: 0.1.0
 */
import type { UserUpdateRole } from './userUpdateRole';
import type { UserUpdateStatus } from './userUpdateStatus';

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  status?: UserUpdateStatus;
  /** @minLength 6 */
  password?: string;
  role?: UserUpdateRole;
}